requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
dj-database-url==1.2.0
Django==3.2.16
django-admin-generator==2.2.1
django-allauth==0.51.0
django-utils2==2.12.1
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-utils==3.4.5
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
whitenoise==6.2.0
